Rename Communicator to ClientCommunicator and added default implementations for ServerCommunicator
